WIP: Add Security Account Manager Class
Class will be used to pull in a file that contains credentials to all
users.  It will than compare the entered username/password against the
file to see if they match.

Needs work on correct implementation.  Right now thinking of using an
array to pull in all username/passwords and then iterate through the
array until a match is found.  If a match is found it will return true
and then use the found index number to compare the hashed password. If
not match is found, it will return false.