Refactor: Username & password must be entered 1st
The user now has to enter the username and password before an incorrect
login error is displayed.  The reason is to prevent information leakage.

If a hacker tries different usernames until one doesn't fail and then
just gets a wrong password error, they then know that user exsits on the
system and can then just work on cracking the password.

With having them enter both and then not knowing what the error is for
prevents this.