Refactor: Remove initializer method
Removed initializer method because setting the objects localized it to
that method and was unable to update the object with other methods. Set
them as a global variable with private settings.

Could have kept initalizer and used 'this' keyword to set the global
variable.  Thought about that after this change.
-    Resolves: issue #16