Update: User files set to variables
Instead of figuring out each user's group and trying to append '.txt' to
the end of it. Put the links to the file in a variable that can now be
called with a simple 'if' statement if that user falls into that user
group.

Also created the stream loop that will print each line of the text file
to the screen.